Fix a compiler warning
git-svn-id: svn://irodssvn.ucsd.edu/trunk@5213 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/lib/core/src/rodsLog.c
@@ -1616,7 +1616,7 @@ generateLogTimestamp(char *ts, int tsLen)
     strftime(timestamp, TIME_LEN, "%Y%m%dT%H%M%S", &utc);
     
     /* 8 characters of '.uuuuuuZ' + nul */
-    if (tsLen < strlen(timestamp) + 9) {
+    if (tsLen < (int)strlen(timestamp) + 9) {
         return;
     }
